app restructure and minor refactoring.
